+++ MindMapBack/MindMapBack/webapp/src/main/java/com/mti/mindmap/persistence/repository/UserMapsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.math.BigInteger;
 
-public interface UserMapsRepository extends CrudRepository<UserMapsModel, BigInteger> {
+public interface UserMapsRepository extends CrudRepository<UserMapsModel, Integer> {
 }
